Include collection and metric in trace details segment names
The first metric in the metrics array will be used to determine the top level
name for the segment in the trace details view.

When 'ActiveRecord/All' is first the segment name will become 'Database' for all
segments. Changing the order so that 'ActiveRecord/collection/metric' is first
will display the more useful 'collection#metric' as the name for each segment.