Do not explicitly metrics - trace_execution_scoped will automatically do it
The previous code was double-counting metrics for all queries through Moped,
because of the fact that trace_execution_scoped will automatically record
metrics. The net effect was that the database band on the overview graph in New
Relic was twice as large as it should have been (because the 'ActiveRecord/all'
metric was being recorded twice).

This change removes the additional explicit call to record_metrics (which is not
a public API anyhow) so that metrics are no longer double-counted.